Accidentally broke file
I still had my files referencing to js/css files within the file named js/css.

I have it this way on my website, although I uploaded these files directly to GitHub on the website, so I could not upload it in the folder. Therefore, I removed that referenced to the folder and fixed some other things regarding getImgData because the function was no longer working.